+++ examples/timer.rs
@@ -2,7 +2,7 @@ use async_std::task;
 use core::time::Duration;
 
 fn main() {
-    pasts::Executor::default().block_on(async {
+    pasts::block_on(async {
         println!("Waiting 2 seconds…");
         task::sleep(Duration::new(2, 0)).await;
         println!("Waited 2 seconds.");